ResolutionStrategyConfigTask and Tests
