+++ dependency-manager/src/functionalTest/resources/bomtests/com/affichage/it21/vk/zentraldispo-dao/9.1.0.ADMIN-UIMIG-SNAPSHOT/_remote.repositories
@@ -1,0 +1,6 @@
+#NOTE: This is a Maven Resolver internal implementation file, its format can be changed without prior notice.
+#Wed Jan 08 09:02:22 CET 2020
+zentraldispo-dao-9.1.0.ADMIN-UIMIG-20191216.095904-223.pom>snapshots=
+zentraldispo-dao-9.1.0.ADMIN-UIMIG-20200107.151250-225.jar>snapshots=
+zentraldispo-dao-9.1.0.ADMIN-UIMIG-20200107.151250-225.pom>snapshots=
+zentraldispo-dao-9.1.0.ADMIN-UIMIG-20191216.095904-223.jar>snapshots=
