+++ dependency-manager/src/functionalTest/resources/bomtests/com/affichage/it21/vk/zentraldispo-dao/9.1.0.ADMIN-UIMIG-SNAPSHOT/maven-metadata-snapshots.xml.sha1
@@ -1,0 +1,1 @@
+9dadebad7869f56d4d81276ad9e206713315b02c
