+++ dependency-manager/src/functionalTest/resources/bomtests/com/affichage/it21/vk/zentraldispo-dao/9.1.0.ADMIN-UIMIG-SNAPSHOT/resolver-status.properties
@@ -1,0 +1,5 @@
+#NOTE: This is a Maven Resolver internal implementation file, its format can be changed without prior notice.
+#Wed Jan 08 09:02:10 CET 2020
+maven-metadata-vaadin-addons.xml.error=
+maven-metadata-snapshots.xml.lastUpdated=1578470530513
+maven-metadata-vaadin-addons.xml.lastUpdated=1578470529744
