+++ dependency-manager/src/functionalTest/resources/bomtests/com/affichage/it21/vk/zentraldispo-dao/9.1.0.ADMIN-UIMIG-SNAPSHOT/zentraldispo-dao-9.1.0.ADMIN-UIMIG-20191216.095904-223.jar.sha1
@@ -1,0 +1,1 @@
+374712897916dc8694f87d6b6e3c00dfec13812f
