+++ dependency-manager/src/functionalTest/resources/bomtests/com/affichage/it21/vk/zentraldispo-dao/9.1.0.ADMIN-UIMIG-SNAPSHOT/zentraldispo-dao-9.1.0.ADMIN-UIMIG-20200107.151250-225.jar.sha1
@@ -1,0 +1,1 @@
+d03686265c4fdab40c4e245a83f12643de777b32
