+++ dependency-manager/src/functionalTest/resources/bomtests/com/affichage/it21/vk/zentraldispo-dao/9.1.0.ADMIN-UIMIG-SNAPSHOT/zentraldispo-dao-9.1.0.ADMIN-UIMIG-20200107.151250-225.pom.sha1
@@ -1,0 +1,1 @@
+90ea2e2055f18b56bb6678c1850dd81420a5b81b
