+++ dependency-manager/src/functionalTest/resources/bomtests/org/apache/httpcomponents/httpclient/4.0.2/_remote.repositories
@@ -1,0 +1,4 @@
+#NOTE: This is a Maven Resolver internal implementation file, its format can be changed without prior notice.
+#Fri Dec 27 13:14:35 CET 2019
+httpclient-4.0.2.jar>central=
+httpclient-4.0.2.pom>central=
