+++ dependency-manager/src/functionalTest/resources/bomtests/org/apache/httpcomponents/httpclient/4.0.2/httpclient-4.0.2.jar.sha1
@@ -1,0 +1,1 @@
+781b68c2fd5335de914166241b8d4bfe8c2f91b7
