+++ dependency-manager/src/functionalTest/resources/bomtests/org/apache/httpcomponents/httpclient/4.0.2/httpclient-4.0.2.pom.sha1
@@ -1,0 +1,1 @@
+f218515f169bb93b7fd6c67fab0dedbaf832a8ae
