+++ dependency-manager/src/functionalTest/resources/bomtests/org/apache/httpcomponents/httpclient/4.5.2/httpclient-4.5.2.jar.sha1
@@ -1,0 +1,1 @@
+733db77aa8d9b2d68015189df76ab06304406e50
