+++ dependency-manager/src/functionalTest/resources/bomtests/org/apache/httpcomponents/httpclient/4.5.2/httpclient-4.5.2.pom.sha1
@@ -1,0 +1,1 @@
+56f6338b324e438307e1f2c2b33bd02268310fc2
