--- dependency-manager/src/main/kotlin/com/apgsga/gradle/maven/dm/ext/DmBombAnalyzersExtension.kt
@@ -1,14 +1,0 @@
-package com.apgsga.gradle.maven.dm.ext
-
-import org.gradle.api.Project
-import org.gradle.api.logging.LogLevel
-
-open class DmBombAnalyzersExtension(private val project : Project) {
-
-    private val boms = listOf<String>()
-
-    fun log() {
-        project.logger.log(LogLevel.INFO,"Boms to analyze: $boms")
-    }
-
-}
