+++ dependency-manager/src/main/resources/META-INF/gradle-plugins/com.apgsga.bom.report.properties
@@ -1,1 +1,1 @@
-implementation-class=com.apgsga.gradle.maven.dm.plugin.DmBombAnalyzerPlugin
+implementation-class=com.apgsga.maven.dm.plugin.DmBombAnalyzerPlugin
