+++ dependency-manager/src/main/resources/META-INF/gradle-plugins/com.apgsga.version.resolver.properties
@@ -1,1 +1,1 @@
-implementation-class=com.apgsga.gradle.maven.dm.plugin.VersionResolutionPlugin
+implementation-class=com.apgsga.maven.dm.plugin.VersionResolutionPlugin
